+++ .vscode/tasks.json
@@ -1,0 +1,34 @@
+{
+    // See https://go.microsoft.com/fwlink/?LinkId=733558
+    // for the documentation about the tasks.json format
+    "version": "2.0.0",
+    "tasks": [
+        {
+            "label": "Run Zomboid Debug No Steam",
+            "type": "shell",
+            "command": "\"E:\\Steam\\steamapps\\common\\ProjectZomboid\\ProjectZomboid64 - nosteam-debug.bat\"",
+            "presentation": {
+                "group": "groupZomboid"
+            },
+            "problemMatcher": [
+                "$eslint-stylish"
+            ]
+        },
+        {
+            "label": "Run two instances of Zomboid Debug No Steam",
+            "dependsOn": [
+                "Run Zomboid Debug No Steam",
+                "Run Zomboid Debug No Steam"
+            ],
+            "problemMatcher": []
+        },
+        {
+            "label": "Run Zomboid Test Server",
+            "type": "shell",
+            "command": "\"E:\\Steam\\steamapps\\common\\Project Zomboid Dedicated Server\\StartServer64_nosteam.bat\"",
+            "problemMatcher": [
+                "$eslint-stylish"
+            ]
+        }
+    ]
+}
