--- contents/mods/AnotherVehicleClaimSystem/media/lua/shared/Translate/EN/IG_UI_EN.txt
@@ -1,3 +1,0 @@
-IG_UI_EN = {
-	IGUI_AVCS_Vehicle_No_Permission = "Not my vehicle"
-}
