--- contents/mods/AnotherVehicleClaimSystem/media/sandbox-options.txt
@@ -1,37 +1,0 @@
-VERSION = 1,
-
-option AVCS.AllowFaction
-{
-		type = boolean, default = true,
-		page = AVCS, translation = AVCS_AllowFaction,
-}
-
-option AVCS.AllowSafehouse
-{
-		type = boolean, default = true,
-		page = AVCS, translation = AVCS_AllowSafehouse,
-}
-
-option AVCS.MaxVehicle
-{
-		type = integer, min = 1, max = 999, default = 5,
-		page = AVCS, translation = AVCS_MaxVehicle,
-}
-
-option AVCS.ReturnTicket
-{
-		type = boolean, default = false,
-		page = AVCS, translation = AVCS_ReturnTicket,
-}
-
-option AVCS.UdderlyRespawn
-{
-		type = boolean, default = false,
-		page = AVCS, translation = AVCS_UdderlyRespawn,
-}
-
-option AVCS.ServerSideChecking
-{
-		type = boolean, default = false,
-		page = AVCS, translation = ServerSideChecking,
-}
