--- contents/mods/AnotherVehicleClaimSystem/mod.info
@@ -1,4 +1,0 @@
-name=Another Vehicle Claim System
-id=AnotherVehicleClaimSystem
-description=Another Vehicle Claim System, it will conflict with other Vehicle Claim mods
-poster=poster.png
