+++ media/lua/shared/Translate/EN/ContextMenu_EN.txt
@@ -1,0 +1,5 @@
+ContextMenu_EN = {
+    ContextMenu_AVCS_ClaimVehicle = "Claim",
+	ContextMenu_AVCS_UnclaimVehicle = "Unclaim",
+    ContextMenu_AVCS_UnsupportedVehicle = "Claim [Error]"
+}
