+++ media/lua/shared/Translate/EN/IG_UI_EN.txt
@@ -1,0 +1,3 @@
+IG_UI_EN = {
+	IGUI_AVCS_Vehicle_No_Permission = "Not my vehicle"
+}
