+++ media/lua/shared/Translate/EN/Sandbox_EN.txt
@@ -1,0 +1,16 @@
+Sandbox_EN = {
+	Sandbox_AVCS = "Another Vehicle Claim System",
+	Sandbox_AVCS_AllowFaction = "Allow Same Faction",
+	Sandbox_AVCS_AllowFaction_tooltip = "Allow members of the same faction to do things to the vehicle",
+	Sandbox_AVCS_AllowSafehouse = "Allow Same Safehouse",
+	Sandbox_AVCS_AllowSafehouse_tooltip = "Allow members of the same safehouse to do things to the vehicle"
+	Sandbox_AVCS_MaxVehicle = "Maximum Vehicle",
+	Sandbox_AVCS_MaxVehicle_tooltip = "Maximum amount of vehicle one can claim",
+	Sandbox_AVCS_ReturnTicket = "Return Claim Ticket on Unclaim",
+	Sandbox_AVCS_ReturnTicket_tooltip = "Claim ticket can be one-time use or can be returned on unclaim",
+	Sandbox_AVCS_UdderlyRespawn = "Udderly Spawn Random Car on Claim",
+	Sandbox_AVCS_UdderlyRespawn_tooltip = "Spawn a random car in the world through Udderly Respawn Mod",
+	Sandbox_ServerSideCheckUnclaim = "Confirm permission on server-side for unclaim",
+	Sandbox_ServerSideCheckUnclaim_tooltip = "This can increase overhead on server-side but you may need it if hackers are rampaging",
+	
+}
