+++ mod.info
@@ -1,0 +1,4 @@
+name=Another Vehicle Claim System
+id=AnotherVehicleClaimSystem
+description=Another Vehicle Claim System, it will conflict with other Vehicle Claim mods
+poster=poster.png
